requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
arrow==1.1.0
    # via
    #   -r requirements.in
    #   django-q
asgiref==3.3.4
    # via django
autopep8==1.5.6
    # via -r requirements.in
blessed==1.18.0
    # via
    #   -r requirements.in
    #   django-q
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
coreapi==2.3.3
    # via -r requirements.in
coreschema==0.0.4
    # via coreapi
django-cors-headers==3.7.0
    # via -r requirements.in
django-picklefield==3.0.1
    # via
    #   -r requirements.in
    #   django-q
django-q==1.3.5
    # via -r requirements.in
django==3.2
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-picklefield
    #   django-q
    #   djangorestframework
djangorestframework==3.12.4
    # via -r requirements.in
gevent==21.1.2
    # via -r requirements.in
greenlet==1.0.0
    # via gevent
gunicorn==20.1.0
    # via -r requirements.in
idna==2.10
    # via requests
itypes==1.2.0
    # via coreapi
jinja2==2.11.3
    # via coreschema
markupsafe==1.1.1
    # via jinja2
pillow==8.2.0
    # via -r requirements.in
psycopg2-binary==2.8.6
    # via -r requirements.in
py3exiv2==0.9.3
    # via -r requirements.in
pycodestyle==2.7.0
    # via autopep8
python-dateutil==2.8.1
    # via arrow
pytz==2021.1
    # via django
pyyaml==5.4.1
    # via -r requirements.in
requests==2.25.1
    # via coreapi
six==1.15.0
    # via
    #   blessed
    #   python-dateutil
sqlparse==0.4.1
    # via django
toml==0.10.2
    # via autopep8
uritemplate==3.0.1
    # via coreapi
urllib3==1.26.4
    # via requests
wcwidth==0.2.5
    # via blessed
zope.event==4.5.0
    # via gevent
zope.interface==5.4.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
